pyproject.toml
[project]
name = "airas-experiment"
version = "0.1.0"
description = "AIRAS ML Experiment Framework"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "openai>=1.12.0",
    "datasets>=2.16.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "scipy>=1.11.0",
]
